refactor: 질문 컴포넌트 분리
